@tinyb0t 2023-07-31
Changes made in `tinygrad/`:
```
------------------------------------------------------------
files                             insertions       deletions
------------------------------------------------------------
tinygrad/helpers.py                        1               0
tinygrad/lazy.py                          24               3
tinygrad/ops.py                            3               3
tinygrad/runtime/lib.py                    8               1
tinygrad/runtime/ops_cuda.py               5               2
tinygrad/runtime/ops_gpu.py                3               2
tinygrad/runtime/ops_metal.py              3               2
------------------------------------------------------------
total                                     47              13
------------------------------------------------------------
lines added in the tinygrad folder: 34
```