@dc-dc-dc 2023-07-31
Can you post the speed results from `test/test_speed_v_torch.py`. Also a before/after of `test/models/test_real_world.py`