@nimlgen 2023-07-31
test_speed_v_torch:
```
add                                1x    1    0.44 ms (    0.00 GFLOPS     0.00 GB/s) in torch,    0.34 ms (    0.00 GFLOPS     0.00 GB/s) in tinygrad,    0.79x faster       0.00 MOPS     0.00 MB
add                             1024x 1024    0.56 ms (    1.88 GFLOPS    22.54 GB/s) in torch,    0.50 ms (    2.09 GFLOPS    25.11 GB/s) in tinygrad,    0.90x faster       1.05 MOPS    12.58 MB
add                             4096x 4096    1.73 ms (    9.72 GFLOPS   116.58 GB/s) in torch,    1.57 ms (   10.66 GFLOPS   127.87 GB/s) in tinygrad,    0.91x faster      16.78 MOPS   201.33 MB
add_constant                    4096x 4096    1.29 ms (   13.02 GFLOPS   104.20 GB/s) in torch,    1.18 ms (   14.23 GFLOPS   113.83 GB/s) in tinygrad,    0.92x faster      16.78 MOPS   134.22 MB
add_sq                          4096x 4096    3.26 ms (   15.46 GFLOPS    61.83 GB/s) in torch,    1.56 ms (   32.24 GFLOPS   128.97 GB/s) in tinygrad,    0.48x faster      50.33 MOPS   201.33 MB
array_packing                   2048x 2048    0.83 ms (    5.05 GFLOPS    40.36 GB/s) in torch,    0.62 ms (    6.75 GFLOPS    54.00 GB/s) in tinygrad,    0.75x faster       4.19 MOPS    33.55 MB
cat_0                            256x  256    0.46 ms (    0.29 GFLOPS     2.30 GB/s) in torch,    0.38 ms (    0.34 GFLOPS     2.74 GB/s) in tinygrad,    0.84x faster       0.13 MOPS     1.05 MB
cat_1                            256x  256    0.44 ms (    0.30 GFLOPS     2.38 GB/s) in torch,    0.40 ms (    0.33 GFLOPS     2.62 GB/s) in tinygrad,    0.91x faster       0.13 MOPS     1.05 MB
conv bs: 32 chans:  4 ->  32 k:3              0.75 ms (  100.64 GFLOPS     6.39 GB/s) in torch,    0.51 ms (  149.19 GFLOPS     9.47 GB/s) in tinygrad,    0.67x faster      75.50 MOPS     4.79 MB
conv bs: 32 chans: 16 ->  32 k:3              0.76 ms (  398.34 GFLOPS     8.68 GB/s) in torch,    0.87 ms (  345.53 GFLOPS     7.53 GB/s) in tinygrad,    1.15x slower     301.99 MOPS     6.58 MB
conv bs: 32 chans: 64 ->  32 k:3              1.30 ms (  929.11 GFLOPS    10.57 GB/s) in torch,    0.93 ms ( 1305.84 GFLOPS    14.85 GB/s) in tinygrad,    0.71x faster    1207.96 MOPS    13.74 MB
double_permute (64, 64, 64, 64)               2.92 ms (    5.75 GFLOPS    46.00 GB/s) in torch,    2.47 ms (    6.79 GFLOPS    54.35 GB/s) in tinygrad,    0.85x faster      16.78 MOPS   134.22 MB
exp                             2048x 2048    0.71 ms (   11.87 GFLOPS    47.46 GB/s) in torch,    0.63 ms (   13.29 GFLOPS    53.15 GB/s) in tinygrad,    0.89x faster       8.39 MOPS    33.55 MB
gemm                            1024x 1024    1.05 ms ( 2043.92 GFLOPS    11.98 GB/s) in torch,    0.94 ms ( 2273.77 GFLOPS    13.32 GB/s) in tinygrad,    0.90x faster    2147.48 MOPS    12.58 MB
gemm                             256x  256    0.46 ms (   72.27 GFLOPS     1.69 GB/s) in torch,    0.43 ms (   77.60 GFLOPS     1.82 GB/s) in tinygrad,    0.93x faster      33.55 MOPS     0.79 MB
gemm_unrolled                    512x  512    0.82 ms (  329.12 GFLOPS     3.86 GB/s) in torch,    0.74 ms (  364.31 GFLOPS     4.27 GB/s) in tinygrad,    0.90x faster     268.44 MOPS     3.15 MB
gemm_unrolled_permute_l          512x  512    0.59 ms (  453.12 GFLOPS     5.31 GB/s) in torch,    0.66 ms (  404.47 GFLOPS     4.74 GB/s) in tinygrad,    1.12x slower     268.44 MOPS     3.15 MB
gemm_unrolled_permute_lr         512x  512    0.55 ms (  488.14 GFLOPS     5.72 GB/s) in torch,    0.63 ms (  427.76 GFLOPS     5.01 GB/s) in tinygrad,    1.14x slower     268.44 MOPS     3.15 MB
gemm_unrolled_permute_r          512x  512    0.75 ms (  358.97 GFLOPS     4.21 GB/s) in torch,    0.45 ms (  596.30 GFLOPS     6.99 GB/s) in tinygrad,    0.60x faster     268.44 MOPS     3.15 MB
max                             4096x 4096    0.98 ms (   17.11 GFLOPS    68.72 GB/s) in torch,    0.88 ms (   19.14 GFLOPS    76.87 GB/s) in tinygrad,    0.89x faster      16.84 MOPS    67.63 MB
mul_sum                         4096x 4096    2.16 ms (   15.56 GFLOPS    62.37 GB/s) in torch,    1.26 ms (   26.78 GFLOPS   107.32 GB/s) in tinygrad,    0.58x faster      33.62 MOPS   134.74 MB
neg                             4096x 4096    1.29 ms (   13.04 GFLOPS   104.30 GB/s) in torch,    1.21 ms (   13.88 GFLOPS   111.05 GB/s) in tinygrad,    0.94x faster      16.78 MOPS   134.22 MB
conv bs:  1 chans: 12 ->  32 k:3              0.55 ms (  103.44 GFLOPS     2.66 GB/s) in torch,    0.87 ms (   65.35 GFLOPS     1.68 GB/s) in tinygrad,    1.58x slower      56.62 MOPS     1.46 MB
partial_sum                     4096x 4096    1.00 ms (   16.82 GFLOPS    67.54 GB/s) in torch,    0.94 ms (   17.96 GFLOPS    72.10 GB/s) in tinygrad,    0.94x faster      16.84 MOPS    67.63 MB
permute                         1024x 1024    0.47 ms (    2.25 GFLOPS    18.00 GB/s) in torch,    0.46 ms (    2.27 GFLOPS    18.18 GB/s) in tinygrad,    0.99x faster       1.05 MOPS     8.39 MB
permute                         4096x 4096    4.46 ms (    3.76 GFLOPS    30.08 GB/s) in torch,    1.20 ms (   13.93 GFLOPS   111.44 GB/s) in tinygrad,    0.27x faster      16.78 MOPS   134.22 MB
pow                             2048x 2048    0.88 ms (  261.56 GFLOPS   114.14 GB/s) in torch,    1.12 ms (  206.73 GFLOPS    90.21 GB/s) in tinygrad,    1.27x slower     230.69 MOPS   100.66 MB
relu                            4096x 4096    1.41 ms (   11.92 GFLOPS    95.38 GB/s) in torch,    1.19 ms (   14.13 GFLOPS   113.01 GB/s) in tinygrad,    0.84x faster      16.78 MOPS   134.22 MB
sub                             4096x 4096    1.68 ms (   10.01 GFLOPS   120.17 GB/s) in torch,    1.57 ms (   10.69 GFLOPS   128.30 GB/s) in tinygrad,    0.94x faster      16.78 MOPS   201.33 MB
sum                             2048x 2048    0.67 ms (    6.25 GFLOPS    25.10 GB/s) in torch,    0.58 ms (    7.30 GFLOPS    29.31 GB/s) in tinygrad,    0.86x faster       4.21 MOPS    16.91 MB
sum                             4096x 4096    0.99 ms (   16.93 GFLOPS    68.00 GB/s) in torch,    0.88 ms (   19.05 GFLOPS    76.51 GB/s) in tinygrad,    0.89x faster      16.84 MOPS    67.63 MB
```

As for `test_real_world.py` my m1pro might run out of memory for llama and sd, maybe better to retest them.
```
New:
METAL=1 python3 models/test_real_world.py
test_llama: used 13.49 GB and 486 kernels in 1601.10 ms
.test_sd: used 14.06 GB and 1016 kernels in 2290.58 ms
.train_cifar: used 7.85 GB and 274 kernels in 1812.34 ms

Old:
METAL=1 python3 models/test_real_world.py
test_llama: used 13.49 GB and 486 kernels in 1633.04 ms
.test_sd: used 14.05 GB and 924 kernels in 2424.88 ms
.train_cifar: used 7.55 GB and 236 kernels in 1841.71 ms
```