@stevenandersonz 2023-07-31
> As for `test_real_world.py` my m1pro might run out of memory for llama and sd, maybe better to retest them.
If you run it with CI=1 it will test the tiny llama