@nimlgen 2023-07-31
> > As for `test_real_world.py` my m1pro might run out of memory for llama and sd, maybe better to retest them.
> 
> If you run it with CI=1 it will test the tiny llama

Thanks. But for llama count of kernels is the same, so the speed should be the same. Llama has reductions with global dims like (242, 64), so it's better not to apply the optimizations, since it just adds more to global/local dims (this amount of threads simply overload gpus, for some gpus performance is unchanged, for some it is worse).

I tested a smaller version of sd (with Tensor(1,4,64,64)->Tensor(1,4,48,48). Overall it's a bit better than original verison:
```
OLD
test_llama: used 13.49 GB and 486 kernels in 1563.78 ms
.test_sd: used 7.30 GB and 924 kernels in 499.63 ms
.train_cifar: used 7.55 GB and 236 kernels in 1830.16 ms

test_llama: used 13.49 GB and 486 kernels in 1617.32 ms
.test_sd: used 7.30 GB and 924 kernels in 501.72 ms
.train_cifar: used 7.55 GB and 236 kernels in 1831.49 ms
  
test_llama: used 13.49 GB and 486 kernels in 1595.49 ms
.test_sd: used 7.30 GB and 924 kernels in 503.38 ms
.train_cifar: used 7.55 GB and 236 kernels in 1826.27 ms

NEW
test_llama: used 13.49 GB and 486 kernels in 1545.04 ms
.test_sd: used 7.30 GB and 928 kernels in 497.91 ms
.train_cifar: used 7.85 GB and 274 kernels in 1786.99 ms

test_llama: used 13.49 GB and 486 kernels in 1537.53 ms
.test_sd: used 7.30 GB and 928 kernels in 498.68 ms
.train_cifar: used 7.85 GB and 274 kernels in 1812.01 ms

test_llama: used 13.49 GB and 486 kernels in 1599.92 ms
.test_sd: used 7.30 GB and 928 kernels in 496.24 ms
.train_cifar: used 7.85 GB and 274 kernels in 1781.60 ms
```