@nimlgen 2023-08-05
Tested this PR on some models from examples/. On what I have tested (see full results for details) this PR is never worse. SD is now even a bit faster after #1429 + this PR.

<details>
<summary>Full results</summary>

```
stable_diffusion.py NEW: avg:  1075.24 GFLOPS    20.06 GB/s           total:  1193 kernels  4221.37 GOPS    78.76 GB  3925.99 ms
stable_diffusion.py OLD: avg:  1085.77 GFLOPS    18.94 GB/s           total:  1076 kernels  4372.68 GOPS    76.29 GB  4027.27 ms

train_resnet.py NEW: 8.04it/s
train_resnet.py OLD: 7.71it/s

efficientnet.py: same
hlb_cifar10.py: same
train_efficientnet.py: same
serious_mnist.py: same
transformer.py: same

test_real_world.py:
  OLD:
  test_sd: used 7.30 GB and 924 kernels in 499.92 ms
  train_cifar (smaller 48x48): used 7.55 GB and 236 kernels in 1834.85 ms
  NEW:
  test_sd: used 7.30 GB and 932 kernels in 495.85 ms
  train_cifar (smaller 48x48): used 7.85 GB and 274 kernels in 1807.73 ms
  test_llama: same
```

</details>